LondonHousePrices/requirements.txt
absl-py==2.0.0
altair==5.1.2
appnope==0.1.3
asttokens==2.4.0
astunparse==1.6.3
attrs==23.1.0
autopep8==2.0.1
backcall==0.2.0
blinker==1.6.3
branca==0.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.7.0
decorator==5.1.1
executing==1.2.0
fiona==1.9.5
flatbuffers==23.5.26
folium==0.15.0
fonttools==4.42.1
future==0.18.3
gast==0.5.4
geopandas==0.14.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mime==0.1.0
missingno==0.5.2
ml-dtypes==0.2.0
mlxtend==0.23.0
nest-asyncio==1.5.7
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.18.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.12.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
streamlit==1.28.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wcwidth==0.2.6
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
